requirements.txt
# Core Libraries
streamlit==1.32.0
PyPDF2==3.0.1
python-dotenv==1.0.1
langchain==0.1.16
langchain-community==0.0.32
langchain-openai==0.1.1
faiss-cpu==1.7.4  # Vector store

# OCR/Image Processing
pytesseract==0.3.10
pillow==10.2.0  # PIL
cobble

# Document Parsing
python-docx==1.1.0
python-pptx==0.6.23
mammoth==1.6.0  # .doc support
pandas==2.2.1  # Excel/CSV handling

# AI/ML
openai==1.12.0  # API client
tiktoken==0.6.0  # Tokenizer
# uncomment to use huggingface llms
# huggingface-hub==0.14.1

# uncomment to use instructor embeddings
# InstructorEmbedding==1.0.1
# sentence-transformers==2.2.2